requirements.txt
absl-py==0.9.0
astor==0.8.1
bleach==3.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
dlib==19.19.0
docutils==0.16
face-recognition==1.3.0
face-recognition-models==0.3.0
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.9
importlib-metadata==1.5.0
jeepney==0.4.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
Markdown==3.2.1
numpy==1.15.4
opencv-python==3.4.9.31
Pillow==7.0.0
pkg-resources==0.0.0
pkginfo==1.5.0.1
protobuf==3.11.3
py-agender==0.0.9
pycparser==2.19
Pygments==2.5.2
PyYAML==5.3
readme-renderer==24.0
requests==2.23.0
requests-toolbelt==0.9.1
scipy==1.4.1
SecretStorage==3.1.2
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.43.0
twine==3.1.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.12.0
zipp==3.0.0
